confirmacao de exclusao para cliente e hardware

+++ app-stocktti-v2/src/app/client/components/client.component.ts
@@ -32,13 +32,16 @@ export class ClientComponent implements OnInit {
   }
 
   deleteById(client_id: number): void {
-    this.clientService.deleteById(client_id).subscribe({
-      next: () => {
-        console.log(`Client with id ${client_id} deleted with sucess. `);
-        this.getClients();
-      },
-      error: err => console.log('Error', err)
-    });
+    const confirma = confirm(`Tem certeza que deseja excluir este cliente?`);
+    if (confirma) {
+      this.clientService.deleteById(client_id).subscribe({
+        next: () => {
+          console.log(`Client with id ${client_id} deleted with sucess. `);
+          this.getClients();
+        },
+        error: err => console.log('Error', err)
+      });
+    }
   }
 
   set filter(value: string) {
